Añadido filtros, modificacion datos perfil

+++ backend/src/main/java/mygamelist/Repositories/JuegoRepository.java
@@ -3,5 +3,5 @@ package mygamelist.repositories;
 import mygamelist.entities.Juego;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JuegoRepository extends JpaRepository<Juego, Long> {
+public interface JuegoRepository extends JpaRepository<Juego, Integer> {
 }
